pr check

+++ Tabloid/Models/UserProfile.cs
@@ -38,7 +38,7 @@ namespace Tabloid.Models
         public int UserTypeId { get; set; }
         public UserType UserType { get; set; }
         public int UserProfileId { get; set; }
-        public int IsDeactivated { get; set; }
+        public byte IsDeactivated { get; set; }
         public string FullName
         {
             get
